Examples

This package provides different tools for optimization. Hence, this section gives different examples for using the implemented Metaheuristics.

Single-Objective Optimization

julia> using Metaheuristics
julia> f(x) = 10length(x) + sum( x.^2 - 10cos.(2π*x) ) # objective functionf (generic function with 1 method)
julia> bounds = [-5ones(10) 5ones(10)]' # limits/bounds2×10 LinearAlgebra.Adjoint{Float64,Array{Float64,2}}:
 -5.0  -5.0  -5.0  -5.0  -5.0  -5.0  -5.0  -5.0  -5.0  -5.0
  5.0   5.0   5.0   5.0   5.0   5.0   5.0   5.0   5.0   5.0
julia> information = Information(f_optimum = 0.0); # information on the minimization problem
julia> options = Options(f_calls_limit = 9000*10, f_tol = 1e-5); # generic settings
julia> algorithm = ECA(information = information, options = options) # metaheuristic used to optimizeECA(η_max=2.0, K=7, N=0, N_init=0, p_exploit=0.95, p_bin=0.02, p_cr=Float64[], ε=0.0, adaptive=false, resize_population=false)
julia> result = optimize(f, bounds, algorithm) # start the minimization process+=========== RESULT ==========+
  iteration: 1286
    minimum: 2.98488
  minimizer: [-8.86378160281256e-10, -2.4295365011559597e-9, -8.445109265452514e-10, 4.826741251231204e-9, -0.9949586382373169, -3.305984672592656e-9, 0.9949586403917353, 2.326282708744905e-9, -3.264317336590925e-10, -0.9949586371128085]
    f calls: 90020
 total time: 2.3998 s
stop reason: Maximum objective function calls exceeded.
+============================+
julia> minimum(result)2.9848771712798623
julia> minimizer(result)10-element Array{Float64,1}:
 -8.86378160281256e-10
 -2.4295365011559597e-9
 -8.445109265452514e-10
  4.826741251231204e-9
 -0.9949586382373169
 -3.305984672592656e-9
  0.9949586403917353
  2.326282708744905e-9
 -3.264317336590925e-10
 -0.9949586371128085
julia> result = optimize(f, bounds, algorithm) # note that second run is faster+=========== RESULT ==========+
  iteration: 1286
    minimum: 2.98488
  minimizer: [-8.86378160281256e-10, -2.4295365011559597e-9, -8.445109265452514e-10, 4.826741251231204e-9, -0.9949586382373169, -3.305984672592656e-9, 0.9949586403917353, 2.326282708744905e-9, -3.264317336590925e-10, -0.9949586371128085]
    f calls: 89950
 total time: 0.3228 s
stop reason: Maximum number of iterations exceeded.
+============================+

Constrained Optimization

It is common that optimization models include constraints that must be satisfied. For example: The Rosenbrock function constrained to a disk

Minimize:

\[{\displaystyle f(x,y)=(1-x)^{2}+100(y-x^{2})^{2}}\]

subject to:

\[{\displaystyle x^{2}+y^{2}\leq 2}\]

where $-2 \leq x,y \leq 2$.

In Metaheuristics.jl, a feasible solution is such that $g(x) \leq 0$ and $h(x) \approx 0$. Hence, in this example the constraint is given by $g(x) = x^2 + y^2 - 2 \leq 0$. Moreover, the equality and inequality constraints must be saved into Arrays.

Constraints handling

In this package, if the algorithm was not designed for constrained optimization, then solutions with the lower constraint violation sum will be preferred.

julia> using Metaheuristics
julia> function f(x)
           x,y = x[1], x[2]
       
           fx = (1-x)^2+100(y-x^2)^2
           gx = [x^2 + y^2 - 2] # inequality constraints
           hx = [0.0] # equality constraints
       
           # order is important
           return fx, gx, hx
       endf (generic function with 1 method)
julia> bounds = [-2.0 -2; 2 2]2×2 Array{Float64,2}:
 -2.0  -2.0
  2.0   2.0
julia> optimize(f, bounds, ECA(N=30, K=3))+=========== RESULT ==========+
  iteration: 667
    minimum: 1.64749e-15
  minimizer: [0.9999999594613886, 0.9999999187200278]
    f calls: 20010
  feasibles: 30 / 30 in final population
 total time: 2.0163 s
stop reason: Maximum objective function calls exceeded.
+============================+

Multiobjective Optimization

To implement a multiobjective optimization problem and solve it, you can proceed as usual. Here, you need to provide constraints if they exist, otherwise put gx = [0.0]; hx = [0.0]; to indicate an unconstrained multiobjective problem.

julia> using Metaheuristics
julia> function f(x)
           # objective functions
           v = 1.0 + sum(x .^ 2)
           fx1 = x[1] * v
           fx2 = (1 - sqrt(x[1])) * v
       
           fx = [fx1, fx2]
       
           # constraints
           gx = [0.0] # inequality constraints
           hx = [0.0] # equality constraints
       
           # order is important
           return fx, gx, hx
       endf (generic function with 1 method)
julia> bounds = [zeros(30) ones(30)]';
julia> optimize(f, bounds, NSGA2())+=========== RESULT ==========+
  iteration: 251
 population:100-element Array{Metaheuristics.xFgh_solution{Array{Float64,1}},1}:
 (f = [5.827959971211975e-7, 0.9993571366013531], g = [0.0], h = [0.0], x = [5.827e-07, 5.934e-03, …, 9.589e-04])
 (f = [2.0001903153519316, 0.0], g = [0.0], h = [0.0], x = [1.000e+00, 3.683e-04, …, 1.898e-03])
 (f = [0.3906912744995246, 0.4608781780749229], g = [0.0], h = [0.0], x = [3.478e-01, 1.073e-02, …, 3.704e-03])
 (f = [1.8214802418352405, 0.04523231233600479], g = [0.0], h = [0.0], x = [9.532e-01, 9.675e-03, …, 9.649e-04])
 (f = [0.4662727654142582, 0.42552436973797847], g = [0.0], h = [0.0], x = [4.015e-01, 5.663e-03, …, 3.552e-03])
 (f = [1.8889100717250964, 0.02787458369366596], g = [0.0], h = [0.0], x = [9.715e-01, 6.188e-03, …, 8.394e-04])
 (f = [0.5621102982261116, 0.38934375167994034], g = [0.0], h = [0.0], x = [4.622e-01, 6.190e-03, …, 3.471e-03])
 (f = [0.5074818313991308, 0.40896123113906735], g = [0.0], h = [0.0], x = [4.286e-01, 5.795e-03, …, 9.467e-04])
 (f = [0.7552130117856768, 0.3295436141066708], g = [0.0], h = [0.0], x = [5.666e-01, 6.054e-03, …, 2.635e-03])
 (f = [0.315734441161823, 0.5002329891613725], g = [0.0], h = [0.0], x = [2.908e-01, 8.913e-04, …, 9.787e-04])
 ⋮
 (f = [0.05066434150082291, 0.7795041739026338], g = [0.0], h = [0.0], x = [5.040e-02, 8.233e-03, …, 9.719e-04])
 (f = [0.1353601155005406, 0.6470720255028451], g = [0.0], h = [0.0], x = [1.329e-01, 8.659e-03, …, 3.657e-03])
 (f = [0.30802115057200247, 0.5048644436391179], g = [0.0], h = [0.0], x = [2.846e-01, 1.180e-02, …, 3.667e-03])
 (f = [0.6034695386323633, 0.37377047751116493], g = [0.0], h = [0.0], x = [4.874e-01, 5.700e-03, …, 1.857e-03])
 (f = [0.5762490849469981, 0.38399113025855885], g = [0.0], h = [0.0], x = [4.709e-01, 6.192e-03, …, 1.705e-03])
 (f = [1.1809341529389374, 0.20711522307089475], g = [0.0], h = [0.0], x = [7.532e-01, 6.164e-03, …, 8.957e-04])
 (f = [0.01193988754469379, 0.8922253570837234], g = [0.0], h = [0.0], x = [1.192e-02, 1.008e-03, …, 1.040e-03])
 (f = [0.14466062741985378, 0.6362973054334996], g = [0.0], h = [0.0], x = [1.418e-01, 3.746e-04, …, 1.092e-04])
 (f = [0.18919406076202572, 0.5919556723720247], g = [0.0], h = [0.0], x = [1.829e-01, 7.022e-04, …, 3.585e-03])
non-dominated solution(s):
100-element Array{Metaheuristics.xFgh_solution{Array{Float64,1}},1}:
 (f = [5.827959971211975e-7, 0.9993571366013531], g = [0.0], h = [0.0], x = [5.827e-07, 5.934e-03, …, 9.589e-04])
 (f = [2.0001903153519316, 0.0], g = [0.0], h = [0.0], x = [1.000e+00, 3.683e-04, …, 1.898e-03])
 (f = [0.3906912744995246, 0.4608781780749229], g = [0.0], h = [0.0], x = [3.478e-01, 1.073e-02, …, 3.704e-03])
 (f = [1.8214802418352405, 0.04523231233600479], g = [0.0], h = [0.0], x = [9.532e-01, 9.675e-03, …, 9.649e-04])
 (f = [0.4662727654142582, 0.42552436973797847], g = [0.0], h = [0.0], x = [4.015e-01, 5.663e-03, …, 3.552e-03])
 (f = [1.8889100717250964, 0.02787458369366596], g = [0.0], h = [0.0], x = [9.715e-01, 6.188e-03, …, 8.394e-04])
 (f = [0.5621102982261116, 0.38934375167994034], g = [0.0], h = [0.0], x = [4.622e-01, 6.190e-03, …, 3.471e-03])
 (f = [0.5074818313991308, 0.40896123113906735], g = [0.0], h = [0.0], x = [4.286e-01, 5.795e-03, …, 9.467e-04])
 (f = [0.7552130117856768, 0.3295436141066708], g = [0.0], h = [0.0], x = [5.666e-01, 6.054e-03, …, 2.635e-03])
 (f = [0.315734441161823, 0.5002329891613725], g = [0.0], h = [0.0], x = [2.908e-01, 8.913e-04, …, 9.787e-04])
 ⋮
 (f = [0.05066434150082291, 0.7795041739026338], g = [0.0], h = [0.0], x = [5.040e-02, 8.233e-03, …, 9.719e-04])
 (f = [0.1353601155005406, 0.6470720255028451], g = [0.0], h = [0.0], x = [1.329e-01, 8.659e-03, …, 3.657e-03])
 (f = [0.30802115057200247, 0.5048644436391179], g = [0.0], h = [0.0], x = [2.846e-01, 1.180e-02, …, 3.667e-03])
 (f = [0.6034695386323633, 0.37377047751116493], g = [0.0], h = [0.0], x = [4.874e-01, 5.700e-03, …, 1.857e-03])
 (f = [0.5762490849469981, 0.38399113025855885], g = [0.0], h = [0.0], x = [4.709e-01, 6.192e-03, …, 1.705e-03])
 (f = [1.1809341529389374, 0.20711522307089475], g = [0.0], h = [0.0], x = [7.532e-01, 6.164e-03, …, 8.957e-04])
 (f = [0.01193988754469379, 0.8922253570837234], g = [0.0], h = [0.0], x = [1.192e-02, 1.008e-03, …, 1.040e-03])
 (f = [0.14466062741985378, 0.6362973054334996], g = [0.0], h = [0.0], x = [1.418e-01, 3.746e-04, …, 1.092e-04])
 (f = [0.18919406076202572, 0.5919556723720247], g = [0.0], h = [0.0], x = [1.829e-01, 7.022e-04, …, 3.585e-03])
    f calls: 50100
  feasibles: 100 / 100 in final population
 total time: 4.9471 s
stop reason: Maximum objective function calls exceeded.
+============================+

Bilevel Optimization

Bilevel optimization problems can be solved by using the package BilevelHeuristics.jl which extends Metaheuristics.jl for handling those hierarchical problems.

Defining objective functions corresponding to the BO problem.

Upper level (leader problem):

using BilevelHeuristics

F(x, y) = sum(x.^2) + sum(y.^2)
bounds_ul = [-ones(5) ones(5)] 

Lower level (follower problem):

f(x, y) = sum((x - y).^2) + y[1]^2
bounds_ll = [-ones(5) ones(5)];

Approximate solution:

res = optimize(F, f, bounds_ul, bounds_ll, BCA())

Output:

+=========== RESULT ==========+
  iteration: 108
    minimum: 
          F: 4.03387e-10
          f: 2.94824e-10
  minimizer: 
          x: [-1.1460768817533927e-5, 7.231706879604178e-6, 3.818596951258517e-6, 2.294324313691869e-6, 1.8770952450067828e-6]
          y: [1.998748659975197e-6, 9.479307908087866e-6, 6.180041276047425e-6, -7.642051857319683e-6, 2.434166021682429e-6]
    F calls: 2503
    f calls: 5062617
    Message: Stopped due UL function evaluations limitations. 
 total time: 26.8142 s
+============================+

See BilevelHeuristics documentation for more information.

Decision-Making

Although Metaheuristics is focused on the optimization part, some decision-making algorithms are available in this package (see Multi-Criteria Decision-Making).

The following example shows how to perform a posteriori decision-making.

julia> # load the problem
julia> f, bounds, pf = Metaheuristics.TestProblems.ZDT1();

julia> # perform multi-objective optimization
julia> res = optimize(f, bounds, NSGA2());

julia> # user preferences
julia> w = [0.5, 0.5];

julia> # set the decision-making algorithm
julia> dm_method = CompromiseProgramming(Tchebysheff())

julia> # find the best decision
julia> sol = best_alternative(res, w, dm_method)
(f = [0.38493217206706115, 0.38037042164979956], g = [0.0], h = [0.0], x = [3.849e-01, 7.731e-06, …, 2.362e-07])

Providing Initial Solutions

Sometimes you may need to use the starter solutions you need before the optimization process begins, well, this example illustrates how to do it.

julia> using Metaheuristics
julia> f, bounds, optimums = Metaheuristics.TestProblems.get_problem(:sphere);
julia> D = size(bounds,2);
julia> x_known = 0.6ones(D) # known solution10-element Array{Float64,1}:
 0.6
 0.6
 0.6
 0.6
 0.6
 0.6
 0.6
 0.6
 0.6
 0.6
julia> X = [ bounds[1,:] + rand(D).* ( bounds[2,:] -  bounds[1,:]) for i in 1:19  ]; # random solutions (uniform distribution)
julia> push!(X, x_known); # save an interest solution
julia> population = [ Metaheuristics.create_child(x, f(x)) for x in X ]; # generate the population with 19+1 solutions
julia> prev_status = State(Metaheuristics.get_best(population), population); # prior state
julia> method = ECA(N = length(population))ECA(η_max=2.0, K=7, N=20, N_init=20, p_exploit=0.95, p_bin=0.02, p_cr=Float64[], ε=0.0, adaptive=false, resize_population=false)
julia> method.status = prev_status; # say to ECA that you have generated a population
julia> optimize(f, bounds, method) # optimize+=========== RESULT ==========+
  iteration: 5001
    minimum: 5.11556e-126
  minimizer: [5.631175827755314e-64, -9.33280723971917e-64, 1.7983488099873907e-64, -8.874212876545405e-64, -2.3628905919419664e-64, -3.91911796241353e-64, -3.686825874167231e-64, 5.251088182792933e-64, -2.293440538935081e-64, 1.5600948799451986e-63]
    f calls: 100000
 total time: 0.6659 s
stop reason: Maximum objective function calls exceeded.
+============================+

Batch Evaluation

Evaluating multiple solutions at the same time can reduce computational time. To do that, define your function on an input N x D matrix and function values into matrices with outcomes in rows for all N solutions. Also, you need to put parallel_evaluation=true in the Options to indicate that your f is prepared for parallel evaluations.

f(X) = begin
    fx = sum(X.^2, dims=2)       # objective function ∑x²
    gx = sum(X.^2, dims=2) .-0.5 # inequality constraints ∑x² ≤ 0.5
    hx = zeros(0,0)              # equality constraints
    fx, gx, hx
end

options = Options(parallel_evaluation=true)

res = optimize(f, [-10ones(5) 10ones(5)], ECA(options=options))

See Parallelization tutorial for more details.

Modifying an Existing Metaheuristic

You may need to modify one of the implemented metaheuristics to improve the algorithm performance or test new mechanisms. This example illustrates how to do it.

Modifying algorithms could break stuff

Be cautious when modifying a metaheuristic due to those changes will overwrite the default method for that metaheuristic.

Let's assume that we want to modify the stop criteria for ECA. See Contributing for more details.

julia> using Metaheuristics
julia> import LinearAlgebra: norm
julia> # overwrite method
       function Metaheuristics.stop_criteria!(
               status,
               parameters::ECA, # It is important to indicate the modified Metaheuristic
               problem,
               information,
               options,
               args...;
               kargs...
           )
       
           if status.stop
               # nothing to do
               return
           end
       
           # Diversity-based stop criteria
       
           x_mean = zeros(length(status.population[1].x))
           for sol in status.population
               x_mean += sol.x
           end
           x_mean /= length(status.population)
       
           distances_mean = sum(sol -> norm( x_mean - sol.x ), status.population)
           distances_mean /= length(status.population)
       
           # stop when solutions are close enough to the geometrical center
           new_stop_condition = distances_mean <= 1e-3
       
           status.stop = new_stop_condition
       
           # (optional and not recommended) print when this criterium is met
           if status.stop
               @info "Diversity-based stop criterium"
               @show distances_mean
           end
       
       
           return
       end
julia> f, bounds, opt = Metaheuristics.TestProblems.get_problem(:sphere);
julia> optimize(f, bounds, ECA())[ Info: Diversity-based stop criterium
distances_mean = 0.0009935090742215602
+=========== RESULT ==========+
  iteration: 179
    minimum: 2.96831e-07
  minimizer: [-0.0001259991993556894, 0.00013747368626084355, 0.0001446712101108381, -0.00010870557318847364, 0.0003315771567829122, -5.4348549731001946e-5, 7.125221003931702e-5, -0.00012564494672053094, 0.00014642275604474287, -0.000272230691096466]
    f calls: 12530
 total time: 0.4103 s
stop reason: Unknown stop reason.
+============================+

Restarting search

Metaheuristics.Restart — Type

Restart(optimizer, every=100)

Resets the optimizer every specified number of iterations (100 by default).

Example

julia> f, bounds, _ = Metaheuristics.TestProblems.rastrigin();

julia> optimize(f, bounds, Restart(ECA(), every=200))

Customization

The restart condition can be updated by overloading the restart_condition method:

function Metaheuristics.restart_condition(status, restart::Restart, information, options)
    st.iteration % params.every == 0
end